+++ pyannote/audio/tasks/separation/__init__.py
@@ -1,6 +1,6 @@
 # MIT License
 #
-# Copyright (c) 2020-2022 CNRS
+# Copyright (c) 2024- CNRS
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -19,19 +19,3 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
-from .multilabel import MultiLabelSegmentation
-from .overlapped_speech_detection import OverlappedSpeechDetection
-from .resegmentation import Resegmentation
-from .speaker_diarization import SpeakerDiarization
-from .speech_separation import SpeechSeparation
-from .voice_activity_detection import VoiceActivityDetection
-
-__all__ = [
-    "VoiceActivityDetection",
-    "OverlappedSpeechDetection",
-    "SpeakerDiarization",
-    "Resegmentation",
-    "MultiLabelSegmentation",
-    "SpeechSeparation",
-]